Fixed the tableau table parsing
html table is turned into a 2D matrix.
Matrix allows us to search through columns of names and pair them up with the score/ref info. Score is contained in the next column to the right of the column with the names of the fencers.

Then we create a list of matches, each match is a list containing [Fencer A, Fencer B, Score and Referee]

Some JS injection was learned to be able to load the whole tableau table into the page all at once.

TODO: add a round header cell that shows which round the match took place in.
TODO: add a winner header cell that shows which of the who fencers won the match.